backend/requirements.txt
fastapi
uvicorn
httpx
google-cloud-bigquery
